+++ packages/markitdown-api/src/markitdown_api/convert_file.py
@@ -4,7 +4,7 @@ from typing import Annotated, Any
 from fastapi import APIRouter, UploadFile, File, Form
 
 from markitdown import StreamInfo
-from markitdown_api.ApiConverter import ApiConverter
+from markitdown_api.api_converter import ApiConverter
 from markitdown_api.api_types import (
     ConvertResult,
     LlmOptions,
