+++ packages/markitdown-api/src/markitdown_api/convert_text.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter
 from pydantic import Field
 
 from markitdown import StreamInfo
-from markitdown_api.ApiConverter import ApiConverter
+from markitdown_api.api_converter import ApiConverter
 from markitdown_api.api_types import (
     ConvertRequest,
     ConvertResult,
