+++ packages/markitdown-api/src/markitdown_api/convert_uri.py
@@ -3,7 +3,7 @@ from typing import Annotated, Any
 from fastapi import Query, Body, APIRouter
 from pydantic import Field
 
-from markitdown_api.ApiConverter import ApiConverter
+from markitdown_api.api_converter import ApiConverter
 from markitdown_api.api_types import (
     ConvertRequest,
     ConvertResult,
